Move transformers to GATKTool
